Restructure to allow usage without built-in locales
